pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bio_lib"
version = "0.1.1"
description = "A library for biological computations with JAX support"
requires-python = ">=3.7"
authors = [
    { name = "mcale6" }
]
dependencies = [
    "biopython",
    "numpy",
    "jax",
    "jaxlib",
]
readme = "README.md"
license = { file = "LICENSE" }

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
run-prodigy = "bio_lib.run_prodigy_jax:main"

[tool.hatch.build.targets.wheel]
packages = ["bio_lib"]
sources = ["src"]

[tool.hatch.build]
artifacts = [
    "src/bio_lib/data/*.xyz",
    "src/bio_lib/data/*.radii"
]